linting and typo fixed
