+++ planet4-gutenberg-blocks.php
@@ -119,7 +119,6 @@ const PAGE_BLOCK_TYPES = [
 	'planet4-blocks/take-action-boxout',
 	'planet4-blocks/timeline',
 	'planet4-blocks/socialshare',
-
 ];
 
 // campaigns allow all block types.
@@ -138,7 +137,6 @@ const CAMPAIGN_BLOCK_TYPES = [
 	'planet4-blocks/submenu',
 	'planet4-blocks/timeline',
 	'planet4-blocks/socialshare',
-
 ];
 
 /**
